+++ components/PromptCardList.tsx
@@ -1,0 +1,16 @@
+import PromptCard from "./PromptCard";
+
+const PromptCardList = ({ data, handleTagClick }) => {
+  return (
+    <div className="mt-16 prompt_layout">
+      {data?.map((post) => (
+        <PromptCard
+          key={post._id}
+          post={post}
+          handleTagClick={handleTagClick}></PromptCard>
+      ))}
+    </div>
+  );
+};
+
+export default PromptCardList;
