--- components/Provider.jsx
@@ -1,9 +1,0 @@
-"use client";
-
-import { SessionProvider } from "next-auth/react";
-
-const Provider = ({ children, session }) => {
-    return <SessionProvider>{children}</SessionProvider>;
-};
-
-export default Provider;
